PEP8 fixes and module renames.
