+++ modoboa/extensions/amavis/app_settings.py
@@ -3,7 +3,7 @@
 from django import forms
 from django.utils.translation import ugettext_lazy as _
 
-from modoboa.lib.formutils import SeparatorField, YesNoField, InlineRadioSelect
+from modoboa.lib.form_utils import SeparatorField, YesNoField, InlineRadioSelect
 from modoboa.lib.parameters import AdminParametersForm, UserParametersForm
