+++ modoboa/extensions/postfix_autoreply/app_settings.py
@@ -2,7 +2,7 @@
 from django.utils.translation import ugettext_lazy
 from django import forms
 from modoboa.lib.parameters import AdminParametersForm
-from modoboa.lib.formutils import SeparatorField
+from modoboa.lib.form_utils import SeparatorField
 
 
 class ParametersForm(AdminParametersForm):
