+++ modoboa/extensions/postfix_relay_domains/app_settings.py
@@ -6,7 +6,7 @@ from django import forms
 from django.utils.translation import ugettext_lazy
 
 from modoboa.lib import parameters
-from modoboa.lib.formutils import SeparatorField
+from modoboa.lib.form_utils import SeparatorField
 
 
 class AdminParametersForm(parameters.AdminParametersForm):
