+++ modoboa/extensions/postfix_relay_domains/forms.py
@@ -6,7 +6,7 @@ from django.http import QueryDict
 from django.utils.translation import ugettext as _, ugettext_lazy
 from django.core.urlresolvers import reverse
 from modoboa.lib import events
-from modoboa.lib.formutils import DynamicForm, DomainNameField, TabForms
+from modoboa.lib.form_utils import DynamicForm, DomainNameField, TabForms
 from modoboa.lib.webutils import render_to_json_response
 from modoboa.extensions.admin.models import Domain, DomainAlias
 from .models import RelayDomain, RelayDomainAlias
