+++ modoboa/extensions/radicale/app_settings.py
@@ -1,7 +1,7 @@
 from django.utils.translation import ugettext_lazy
 from django import forms
 from modoboa.lib.parameters import AdminParametersForm
-from modoboa.lib.formutils import SeparatorField, YesNoField
+from modoboa.lib.form_utils import SeparatorField, YesNoField
 
 
 class ParametersForm(AdminParametersForm):
