+++ modoboa/extensions/webmail/app_settings.py
@@ -2,7 +2,7 @@
 from django import forms
 from django.utils.translation import ugettext_lazy as _
 from modoboa.lib.parameters import AdminParametersForm, UserParametersForm
-from modoboa.lib.formutils import SeparatorField, YesNoField, InlineRadioSelect
+from modoboa.lib.form_utils import SeparatorField, YesNoField, InlineRadioSelect
 
 
 class ParametersForm(AdminParametersForm):
